+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # [unreleased]
 
-Breaking changes
+Breaking changes:
 
 * Removed diesel integration. If you were using it, please comment on the corresponding issue:
   https://github.com/ruma/ruma-identifiers/issues/22
@@ -8,6 +8,10 @@ Breaking changes
 * Update `parse_with_server_name`s signature (instead of `Into<String>` it now requires
   `Into<Box<str>>` of the id type). This is technically a breaking change, but extremely unlikely
   to affect any existing code.
+
+Improvements:
+
+* Add `RoomVersionId::version_6` and `RoomVersionId::is_version_6`
 
 # 0.16.1